+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * `--upstream-tag-ends` split into `--upstream-tag-biggest-end` &
   `--upstream-tag-ends-full`. `biggest` will only include data for the end with
   the largest upstream value. This uses a little less memory.
+* Remove Strahler number calculation. It doesn't work well for OSM data
 
 ## v0.29.0 (2024-05-27)